+++ core/src/test/java/bisq/core/trade/validation/DepositTxValidationTest.java
@@ -20,20 +20,8 @@ package bisq.core.trade.validation;
 import bisq.core.btc.model.RawTransactionInput;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.TradeWalletService;
-import bisq.core.dao.burningman.DelayedPayoutTxReceiverService;
 import bisq.core.offer.Offer;
-import bisq.core.provider.fee.FeeService;
-import bisq.core.provider.price.PriceFeedService;
-import bisq.core.trade.TradeFeeFactory;
 import bisq.core.trade.model.bisq_v1.Trade;
-import bisq.core.trade.protocol.bisq_v1.messages.InputsForDepositTxRequest;
-import bisq.core.user.User;
-
-import bisq.network.p2p.NodeAddress;
-
-import bisq.common.crypto.CryptoException;
-import bisq.common.crypto.PubKeyRing;
-import bisq.common.crypto.Sig;
 
 import com.google.protobuf.ByteString;
 
@@ -47,13 +35,8 @@ import org.bitcoinj.core.Transaction;
 import org.bitcoinj.core.TransactionInput;
 import org.bitcoinj.core.TransactionOutPoint;
 import org.bitcoinj.core.TransactionWitness;
-import org.bitcoinj.params.MainNetParams;
 import org.bitcoinj.script.Script;
 import org.bitcoinj.script.ScriptBuilder;
-
-import java.security.KeyPair;
-
-import java.nio.charset.StandardCharsets;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -62,8 +45,8 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import static bisq.core.trade.validation.DepositTxValidation.checkDepositTxMatchesIgnoringWitnessesAndScriptSigs;
-import static bisq.core.trade.validation.TradeAmountValidationTest.OFFER_MIN_AMOUNT;
-import static bisq.core.trade.validation.ValidationTestUtils.*;
+import static bisq.core.trade.validation.ValidationTestUtils.PARAMS;
+import static bisq.core.trade.validation.ValidationTestUtils.btcWalletService;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
@@ -72,12 +55,16 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class DepositTxValidationTest {
+class DepositTxValidationTest {
     static final String SELLER_ADDRESS = SegwitAddress.fromKey(PARAMS, new ECKey()).toString();
     static final String BUYER_ADDRESS = SegwitAddress.fromKey(PARAMS, new ECKey()).toString();
 
     private static final String MAKER_ROLE = "Maker";
     private static final String TAKER_ROLE = "Taker";
+
+    /* --------------------------------------------------------------------- */
+    // Deposit tx comparison
+    /* --------------------------------------------------------------------- */
 
     @Test
     void checkDepositTxMatchesIgnoringWitnessesAndScriptSigsAcceptsMatchingTxsAndReturnsDepositTx() {
@@ -144,6 +131,10 @@ public class DepositTxValidationTest {
                         (BtcWalletService) null));
     }
 
+    /* --------------------------------------------------------------------- */
+    // Unsigned transaction
+    /* --------------------------------------------------------------------- */
+
     @Test
     void checkTransactionIsUnsignedAcceptsValidUnsignedTransaction() {
         byte[] depositTxWithoutWitnesses = ValidationTestUtils.serializedTransaction();
@@ -193,6 +184,10 @@ public class DepositTxValidationTest {
                 ValidationTestUtils.serializedTransaction(),
                 null));
     }
+
+    /* --------------------------------------------------------------------- */
+    // Maker and taker inputs
+    /* --------------------------------------------------------------------- */
 
     @Test
     void checkTakersRawTransactionInputsAcceptsSellerInputsForBuyOffer() {
@@ -303,6 +298,10 @@ public class DepositTxValidationTest {
                 () -> DepositTxValidation.checkRawTransactionInputsAreNotMalleable(List.of(ValidationTestUtils.rawTransactionInput(Coin.SATOSHI)),
                         null));
     }
+
+    /* --------------------------------------------------------------------- */
+    // Peer input validation
+    /* --------------------------------------------------------------------- */
 
     @Test
     void acceptsExactExpectedInputAmountForP2WHInputs() {
@@ -501,100 +500,4 @@ public class DepositTxValidationTest {
     static Transaction copy(Transaction transaction) {
         return new Transaction(PARAMS, transaction.bitcoinSerialize());
     }
-
-
-    static InputsForDepositTxRequestValidationFixture inputsForDepositTxRequestValidationFixture(
-            byte[] accountAgeWitnessSignatureOverride) throws CryptoException {
-        return inputsForDepositTxRequestValidationFixture(accountAgeWitnessSignatureOverride, Coin.valueOf(100));
-    }
-
-    static InputsForDepositTxRequestValidationFixture inputsForDepositTxRequestValidationFixture(
-            byte[] accountAgeWitnessSignatureOverride,
-            Coin requestTakerFee) throws CryptoException {
-        String offerId = "offer-id";
-        Coin tradeAmount = Coin.valueOf(3_000);
-        Coin expectedTakerFee = Coin.valueOf(100);
-        FeeService feeService = configureTradeFeeService(Coin.valueOf(77), expectedTakerFee, 2);
-        Coin tradeTxFee = TradeFeeFactory.getTradeTxFee(feeService.getTxFeePerVbyte());
-        NodeAddress mediatorNodeAddress = new NodeAddress("mediator.onion", 9999);
-        BtcWalletService btcWalletService = btcWalletService();
-        Offer offer = offer(true, Coin.valueOf(10_000), Coin.valueOf(12_000), Coin.valueOf(40_000));
-        when(offer.getId()).thenReturn(offerId);
-        when(offer.getMinAmount()).thenReturn(OFFER_MIN_AMOUNT);
-        when(offer.isUseMarketBasedPrice()).thenReturn(true);
-
-        KeyPair takerSignatureKeyPair = Sig.generateKeyPair();
-        PubKeyRing takerPubKeyRing = pubKeyRing(takerSignatureKeyPair);
-        byte[] accountAgeWitnessSignature = accountAgeWitnessSignatureOverride != null ?
-                accountAgeWitnessSignatureOverride :
-                Sig.sign(takerSignatureKeyPair.getPrivate(), offerId.getBytes(StandardCharsets.UTF_8));
-
-        List<RawTransactionInput> rawTransactionInputs = rawTransactionInputs(btcWalletService,
-                offer.getSellerSecurityDeposit()
-                        .add(tradeAmount)
-                        .add(tradeTxFee.multiply(2)));
-        InputsForDepositTxRequest request = new InputsForDepositTxRequest(offerId,
-                new NodeAddress("sender.onion", 9999),
-                tradeAmount.value,
-                50_000_000L,
-                tradeTxFee.value,
-                requestTakerFee.value,
-                true,
-                rawTransactionInputs,
-                new ECKey().getPubKey(),
-                SegwitAddress.fromKey(MainNetParams.get(), new ECKey()).toString(),
-                takerPubKeyRing,
-                "taker-account-id",
-                ValidationTestUtils.VALID_TRANSACTION_ID,
-                List.of(),
-                List.of(mediatorNodeAddress),
-                List.of(),
-                null,
-                mediatorNodeAddress,
-                new NodeAddress("refund-agent.onion", 9999),
-                "uid",
-                1,
-                accountAgeWitnessSignature,
-                System.currentTimeMillis(),
-                new byte[]{2},
-                "SEPA",
-                130);
-        User user = userWithAcceptedMediator(mediatorNodeAddress,
-                mediator(mediatorNodeAddress, pubKeyRing(Sig.generateKeyPair())));
-
-        return new InputsForDepositTxRequestValidationFixture(request,
-                offer,
-                user,
-                btcWalletService,
-                mock(PriceFeedService.class),
-                delayedPayoutTxReceiverService(130),
-                feeService);
-    }
-
-    static class InputsForDepositTxRequestValidationFixture {
-        final InputsForDepositTxRequest request;
-        final Offer offer;
-        final User user;
-        final BtcWalletService btcWalletService;
-        final PriceFeedService priceFeedService;
-        final DelayedPayoutTxReceiverService delayedPayoutTxReceiverService;
-        final FeeService feeService;
-
-        InputsForDepositTxRequestValidationFixture(InputsForDepositTxRequest request,
-                                                   Offer offer,
-                                                   User user,
-                                                   BtcWalletService btcWalletService,
-                                                   PriceFeedService priceFeedService,
-                                                   DelayedPayoutTxReceiverService delayedPayoutTxReceiverService,
-                                                   FeeService feeService) {
-            this.request = request;
-            this.offer = offer;
-            this.user = user;
-            this.btcWalletService = btcWalletService;
-            this.priceFeedService = priceFeedService;
-            this.delayedPayoutTxReceiverService = delayedPayoutTxReceiverService;
-            this.feeService = feeService;
-        }
-    }
-
 }
