+++ core/src/test/java/bisq/core/trade/validation/TradePriceValidationTest.java
@@ -30,7 +30,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class TradePriceValidationTest {
+class TradePriceValidationTest {
 
     @Test
     void checkTakersTradePriceAcceptsVerifiedPrice() throws Exception {
