+++ core/src/test/java/bisq/core/trade/validation/ValidationTestUtils.java
@@ -26,18 +26,24 @@ import bisq.core.dao.state.DaoStateService;
 import bisq.core.filter.FilterManager;
 import bisq.core.offer.Offer;
 import bisq.core.provider.fee.FeeService;
+import bisq.core.provider.price.PriceFeedService;
 import bisq.core.support.dispute.mediation.mediator.Mediator;
+import bisq.core.trade.TradeFeeFactory;
 import bisq.core.trade.model.bisq_v1.Trade;
 import bisq.core.trade.protocol.TradeMessage;
+import bisq.core.trade.protocol.bisq_v1.messages.InputsForDepositTxRequest;
 import bisq.core.user.User;
 
 import bisq.network.p2p.NodeAddress;
 
+import bisq.common.crypto.CryptoException;
 import bisq.common.crypto.Encryption;
 import bisq.common.crypto.PubKeyRing;
+import bisq.common.crypto.Sig;
 
 import org.bitcoinj.core.Coin;
 import org.bitcoinj.core.ECKey;
+import org.bitcoinj.core.SegwitAddress;
 import org.bitcoinj.core.Sha256Hash;
 import org.bitcoinj.core.Transaction;
 import org.bitcoinj.core.TransactionInput;
@@ -47,6 +53,8 @@ import org.bitcoinj.params.MainNetParams;
 import org.bitcoinj.script.ScriptBuilder;
 
 import java.security.KeyPair;
+
+import java.nio.charset.StandardCharsets;
 
 import java.math.BigInteger;
 
@@ -61,6 +69,9 @@ class ValidationTestUtils {
     static final String VALID_TRANSACTION_ID =
             "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef";
 
+    /* --------------------------------------------------------------------- */
+    // Services and keys
+    /* --------------------------------------------------------------------- */
 
     static BtcWalletService btcWalletService() {
         BtcWalletService btcWalletService = mock(BtcWalletService.class);
@@ -96,6 +107,10 @@ class ValidationTestUtils {
         feeService.updateFeeInfo(txFeePerVbyte, 1);
         return feeService;
     }
+
+    /* --------------------------------------------------------------------- */
+    // Trade domain fixtures
+    /* --------------------------------------------------------------------- */
 
     static Offer offer(boolean isBuyOffer,
                        Coin buyerSecurityDeposit,
@@ -139,6 +154,10 @@ class ValidationTestUtils {
         when(tradeMessage.getTradeId()).thenReturn(tradeId);
         return tradeMessage;
     }
+
+    /* --------------------------------------------------------------------- */
+    // Transaction fixtures
+    /* --------------------------------------------------------------------- */
 
     static byte[] serializedTransaction() {
         return transaction(new byte[]{}).bitcoinSerialize();
@@ -214,5 +233,103 @@ class ValidationTestUtils {
         return new RawTransactionInput(0,
                 parentTransaction,
                 inputAmount.value);
+    }
+
+    /* --------------------------------------------------------------------- */
+    // InputsForDepositTxRequest fixture
+    /* --------------------------------------------------------------------- */
+
+    static InputsForDepositTxRequestValidationFixture inputsForDepositTxRequestValidationFixture(
+            byte[] accountAgeWitnessSignatureOverride) throws CryptoException {
+        return inputsForDepositTxRequestValidationFixture(accountAgeWitnessSignatureOverride, Coin.valueOf(100));
+    }
+
+    static InputsForDepositTxRequestValidationFixture inputsForDepositTxRequestValidationFixture(
+            byte[] accountAgeWitnessSignatureOverride,
+            Coin requestTakerFee) throws CryptoException {
+        String offerId = "offer-id";
+        Coin tradeAmount = Coin.valueOf(3_000);
+        Coin expectedTakerFee = Coin.valueOf(100);
+        FeeService feeService = configureTradeFeeService(Coin.valueOf(77), expectedTakerFee, 2);
+        Coin tradeTxFee = TradeFeeFactory.getTradeTxFee(feeService.getTxFeePerVbyte());
+        NodeAddress mediatorNodeAddress = new NodeAddress("mediator.onion", 9999);
+        BtcWalletService btcWalletService = btcWalletService();
+        Offer offer = offer(true, Coin.valueOf(10_000), Coin.valueOf(12_000), Coin.valueOf(40_000));
+        when(offer.getId()).thenReturn(offerId);
+        when(offer.getMinAmount()).thenReturn(Coin.valueOf(1_000));
+        when(offer.isUseMarketBasedPrice()).thenReturn(true);
+
+        KeyPair takerSignatureKeyPair = Sig.generateKeyPair();
+        PubKeyRing takerPubKeyRing = pubKeyRing(takerSignatureKeyPair);
+        byte[] accountAgeWitnessSignature = accountAgeWitnessSignatureOverride != null ?
+                accountAgeWitnessSignatureOverride :
+                Sig.sign(takerSignatureKeyPair.getPrivate(), offerId.getBytes(StandardCharsets.UTF_8));
+
+        List<RawTransactionInput> rawTransactionInputs = rawTransactionInputs(btcWalletService,
+                offer.getSellerSecurityDeposit()
+                        .add(tradeAmount)
+                        .add(tradeTxFee.multiply(2)));
+        InputsForDepositTxRequest request = new InputsForDepositTxRequest(offerId,
+                new NodeAddress("sender.onion", 9999),
+                tradeAmount.value,
+                50_000_000L,
+                tradeTxFee.value,
+                requestTakerFee.value,
+                true,
+                rawTransactionInputs,
+                new ECKey().getPubKey(),
+                SegwitAddress.fromKey(MainNetParams.get(), new ECKey()).toString(),
+                takerPubKeyRing,
+                "taker-account-id",
+                VALID_TRANSACTION_ID,
+                List.of(),
+                List.of(mediatorNodeAddress),
+                List.of(),
+                null,
+                mediatorNodeAddress,
+                new NodeAddress("refund-agent.onion", 9999),
+                "uid",
+                1,
+                accountAgeWitnessSignature,
+                System.currentTimeMillis(),
+                new byte[]{2},
+                "SEPA",
+                130);
+        User user = userWithAcceptedMediator(mediatorNodeAddress,
+                mediator(mediatorNodeAddress, pubKeyRing(Sig.generateKeyPair())));
+
+        return new InputsForDepositTxRequestValidationFixture(request,
+                offer,
+                user,
+                btcWalletService,
+                mock(PriceFeedService.class),
+                delayedPayoutTxReceiverService(130),
+                feeService);
+    }
+
+    static final class InputsForDepositTxRequestValidationFixture {
+        final InputsForDepositTxRequest request;
+        final Offer offer;
+        final User user;
+        final BtcWalletService btcWalletService;
+        final PriceFeedService priceFeedService;
+        final DelayedPayoutTxReceiverService delayedPayoutTxReceiverService;
+        final FeeService feeService;
+
+        InputsForDepositTxRequestValidationFixture(InputsForDepositTxRequest request,
+                                                   Offer offer,
+                                                   User user,
+                                                   BtcWalletService btcWalletService,
+                                                   PriceFeedService priceFeedService,
+                                                   DelayedPayoutTxReceiverService delayedPayoutTxReceiverService,
+                                                   FeeService feeService) {
+            this.request = request;
+            this.offer = offer;
+            this.user = user;
+            this.btcWalletService = btcWalletService;
+            this.priceFeedService = priceFeedService;
+            this.delayedPayoutTxReceiverService = delayedPayoutTxReceiverService;
+            this.feeService = feeService;
+        }
     }
 }
